Add functionality for adding weather emojis to the forecast and moving utility functions from ForecastScreen to utility js file
