+++ src/screen/ForecastScreen.jsx
@@ -1,47 +1,17 @@
 import { useEffect } from "react";
 import { userForecast } from "../store/index.js";
-import { formatDate, formatTime, cardColor } from "../utils/utils.js";
+import {
+  formatDate,
+  cardColor,
+  getWeatherIcon,
+  getWeeklyForecast,
+  getTodaysForecast,
+  getHourlyForecast,
+} from "../utils/utils.js";
 import { Card } from "@/components/ui/card.jsx";
 import HourlyForecastChart from "@/component/HourlyTempChart.jsx";
 import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
 import { Grid } from "@/component/Grid.jsx";
-
-const getWeeklyForecast = (weeklyData) => {
-  const [...weeklyForecast] = weeklyData.properties.periods;
-  return {
-    forecast: weeklyForecast,
-  };
-};
-
-const getTodaysForecast = (weeklyData) => {
-  const today = weeklyData.properties.periods[0];
-  const todaysTemperature = today.temperature;
-  const [todaysTempColor, todaysTempTextColor] = cardColor(today.temperature);
-  return {
-    forecast: today,
-    temperature: `${todaysTemperature}F`,
-    tempColor: todaysTempColor,
-    tempTextColor: todaysTempTextColor,
-  };
-};
-
-const getHourlyForecast = (hourlyData) => {
-  const arr = [];
-  const periods = hourlyData.properties.periods.slice(0, 19);
-  for (const {
-    startTime,
-    temperature,
-    probabilityOfPrecipitation: { value: percentage },
-  } of periods) {
-    const obj = {
-      time: formatTime(startTime),
-      temperature: temperature,
-      precipitation: percentage,
-    };
-    arr.push(obj);
-  }
-  return arr;
-};
 
 const ForecastScreen = ({ header, props, msgs }) => {
   const { setData, forecastDataArray, setLoading, isLoading } = userForecast();
@@ -113,6 +83,7 @@ const ForecastScreen = ({ header, props, msgs }) => {
               alt={today.forecast.shortForecast}
             ></img>
             Wind: {today.forecast.windSpeed}
+            {getWeatherIcon(today.forecast.shortForecast)}
             {today.tempColor}
           </Grid.Left>
           <Grid.Right>
@@ -130,6 +101,7 @@ const ForecastScreen = ({ header, props, msgs }) => {
                       <Card className={`min-w-80 ${bgColor} ${textColor}`}>
                         <h3>{day.temperature}</h3>
                         <p>{day.shortForecast}</p>
+                        <p>{getWeatherIcon(day.shortForecast)}</p>
                         <p>
                           {day.name} {formatDate(day.startTime)}
                         </p>
